fixup! Reformat code and capitalize script name
