+++ godot/Scenes/characters/player.tscn
@@ -1,6 +1,6 @@
 [gd_scene load_steps=4 format=3 uid="uid://b2jd7hdaj6jik"]
 
-[ext_resource type="Script" path="res://Scenes/characters/player.cs" id="1_42mkc"]
+[ext_resource type="Script" path="res://Scenes/characters/Player.cs" id="1_ywj68"]
 
 [sub_resource type="CylinderShape3D" id="CylinderShape3D_w8hu6"]
 
@@ -8,7 +8,7 @@
 
 [node name="Player" type="CharacterBody3D"]
 transform = Transform3D(1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 1, 0)
-script = ExtResource("1_42mkc")
+script = ExtResource("1_ywj68")
 
 [node name="CameraPivot" type="Node3D" parent="."]
 transform = Transform3D(1, 0, 0, 0, 1, 0, 0, 0, 1, 0, 0.469669, 0)
